pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "openai>=2.14.0",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
